+++ LibraryLoans.Core/Exceptions/NotFoundRestException.cs
@@ -5,5 +5,5 @@ namespace LibraryLoans.Core.Exceptions;
 
 public class NotFoundRestException : BaseRestException
 {
-    public NotFoundRestException(string message) : base((int) HttpStatusCode.NotFound, message) { }
+    public NotFoundRestException(string message) : base(HttpStatusCode.NotFound, message) { }
 }
